refactor: add jest
