+++ jest.config.js
@@ -1,0 +1,8 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+// Sync object
+const config = {
+    verbose: true
+};
+exports.default = config;
+//# sourceMappingURL=jest.config.js.map
