+++ package.json
@@ -7,18 +7,21 @@
     "start": "npm run build && node dist/server.js",
     "build": "rimraf ./dist && npx tsc",
     "lint": "eslint . --ext .ts",
-    "watch": "npx tsc --watch"
+    "watch": "npx tsc --watch",
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
     "@types/express": "^4.17.13",
+    "@types/jest": "^27.0.3",
     "@types/mysql": "^2.15.19",
     "@types/node": "^16.10.3",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "^4.33.0",
     "eslint": "^7.32.0",
+    "jest": "^27.4.3",
     "rimraf": "^3.0.2",
     "ts-node": "^10.4.0",
     "typescript": "^4.4.4"
