+++ src/controllers/tests/AuthController.unit.spec.ts
@@ -1,0 +1,5 @@
+describe("Temp testing example", () => {
+    test("Test jest", () => {
+        expect(() => true);
+    });
+});
